Adder feature added using TDD
