+++ src/main/java/com/qa/calculator/Adder.java
@@ -1,0 +1,10 @@
+package com.qa.calculator;
+
+public class Adder {
+
+	public int add(int num1, int num2) {
+		
+		return num1 + num2;
+	}
+
+}
